feat(src/app.vue src/components/learnvue/index.vue): adjust learnvue content into a component
