--- src/App.css
@@ -1,10 +1,0 @@
-#app {
-    position: absolute;
-    height: 100%;
-    width: 100%;
-    font-family: Avenir, Helvetica, Arial, sans-serif;
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-    text-align: center;
-    color: #2c3e50;
-  }
